Add files via upload
This is my html code for project. In this project we made e-commerce website which aim is to give whole information about company means how they for work, in which field they work, services they offer, industry they serve, event that happen in company, and career opportunities they have and in service they have mobile app development, web & CMS development, game app development, e-commerce development, salesforce solution, AI development, DevOps, Blockchain development, IoT and Embedded like this and in industry there are many industry they serve in every field and in this we also give app and there work in different page after that there is contectus page which help to contect them and yes in this page we have address of company with map. Is an request to see my work on laptop because I'm working in this to make it responsive as early as possible.